+++ src/PharBuild/Builder.php
@@ -382,6 +382,7 @@ class Builder
         printf("api......: %s\n", $pinfo::apiVersion());
         printf("extension: %s\n", phpversion('phar'));
         printf("php......: %s\n", PHP_VERSION);
+        printf("composer.: %s\n", exec('composer -n --version 2>/dev/null'));
         printf("uname....: %s\n", php_uname('a'));
         printf("count....: %d file(s)\n", $pinfo->count());
         $sig = $pinfo->getSignature();