+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to Pipelines will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Changed
+- Pin composer version in phar build; show which composer version in use
+
 ## [0.0.28] - 2020-03-16
 ### Changed
 - Phar build: Do not require platform dependencies for composer install