+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to Pipelines will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Fixed
+- Correct missing link in Change Log
+
 ## [0.0.30] - 2020-04-05
 ### Added
 - Support for `after-script:` incl. `BITBUCKET_EXIT_CODE` environment parameter.
@@ -284,4 +288,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [0.0.27]: https://github.com/ktomk/pipelines/releases/tag/0.0.27
 [0.0.28]: https://github.com/ktomk/pipelines/releases/tag/0.0.28
 [0.0.29]: https://github.com/ktomk/pipelines/releases/tag/0.0.29
+[0.0.30]: https://github.com/ktomk/pipelines/releases/tag/0.0.30
 [unreleased]: https://github.com/ktomk/pipelines